New Piecewise syntax and fix Piecewise issues
Piecewise syntax is changed to allow for a more sensible definition of
the otherwise condition. Before, an expr/cond pair was specified where
cond was given as True. Now, otherwise can either be passed as the final
argument after a sequence of expr/cond 2-tuples. The old otherwise
syntax using (otherwise, True) raises a deprecation warning.

The ExprCondPair class is removed in favor of Tuples to store the
expr/cond pairs.

The '.as_leading_term()' function is changed so that it takes the sum of
all the expressions and, if it is not NaN, the otherwise statement, and
returns the leading term of the sum. Ideally, '.as_leading_term()' would
return a Piecewise function, but doing this breaks the evaluation of
'abs(x-a).nseries(x, 1)'.

Piecewise no longer accepts an Interval as a condition. To implement the
same functionality, either the '.as_relational()' or '.contains()'
methods of Interval can be used. Piecewise also no longer accepts
numbers as conditions.

The properties '.exprcondpairs' and '.otherwise' are implemented to
easily access the expr/cond pairs and otherwise statement, respectively.

Addresses issues: 2101, 2567, 2626, 2710, 2726