Additional Piecewise testing
Changes allowing evaluate keyword to override default behavior of
clearing nested Piecewise expressions. More testing of new Piecewise
evaluate keyword behavior, including an XFAIL test related to Issue
3025, where evaluate=False does not override behavior when bools are in
the conditions.

Includes new tests and doctests for '.exprcondpairs' and '.otherwise'
properties.

Re-enables testing for expressions in expr/cond pairs and the otherwise
statement to be Expr or GeometryEntity. GeometryEntity is needed due to
the use of Piecewise in the geometry module for '.arbitrary_point()'.