Change Piecewise evaluation behavior
Piecewise evaluation now returns the first expression which is
explicitly True. If no conditions are True, the otherwise expression is
returned. Evaluation can be triggered by setting the 'evaluate' keyword
to True. When a boolean is given as a condition, evaluation is enabled
unless specifically disabled by setting 'evaluate' to False. In this
case, any explicitly False conditions are removed and any explicitly
True conditions are given as the otherwise statement. This is done so
bools are kept out of args, see Issue 3025. The collapsing of nested
Piecewise functions is moved from eval to the __new__ method. This is
performed by default, but setting 'evaluate' to False will disable it.