Update init_printing IPython latex formatters
The text/latex representation is now enabled by default.

The `use_latex` option is changed to `png_latex`, and it now controls
how LaTeX is rendered to png. The options are `external`, which uses an
external LaTeX rendering (formerly `png`), and `matplotlib`, which uses
matplotlib to do LaTeX rendering (same as before).